Move IteratesGradeAttributes to shared services
