+++ app/services/shared/iterates_grade_attributes.rb
@@ -4,7 +4,7 @@ module Services
   module Actions
     class IteratesGradeAttributes
       extend LightService::Action
-      
+
       expects :assignment_id, :graded_by_id, :grade_attributes
       promises :successful, :unsuccessful
